Refactored appscale-bootstrap.
Removed downloader wrapper script and instead upload the actual script
directly to the host being imaged.

Also removed the logic for locating the downloader script to use and
simply provide the bootstrap script.

This also looks to find the bootstrap script in a system directory if it
is not being run from a repository.